moved metadata related code to metadata file and created metadata object
